data_collection/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
openaq==0.4.0